requirements.txt
annotated-types      ==0.7.0
anyio                ==4.11.0
certifi              ==2025.11.12
charset-normalizer   ==3.4.4
colorama             ==0.4.6
distro               ==1.9.0
fsspec               ==2024.12.0
h11                  ==0.16.0
httpcore             ==1.0.9
httpx                ==0.28.1
idna                 ==3.11
jiter                ==0.12.0
json_repair          ==0.44.1
jsonpatch            ==1.33
jsonpointer          ==3.0.0
langchain            ==1.1.0
langchain-core       ==1.1.0
langchain-openai     ==1.1.0
langgraph            ==1.0.4
langgraph-checkpoint ==3.0.1
langgraph-prebuilt   ==1.0.5
langgraph-sdk        ==0.2.10
langsmith            ==0.4.48
neo4j                ==5.28.2
neo4j-graphrag       ==1.10.1
numpy                ==2.3.5
openai               ==2.8.1
orjson               ==3.11.4
ormsgpack            ==1.12.0
packaging            ==25.0
pip                  ==25.3
pydantic             ==2.12.4
pydantic_core        ==2.41.5
pypdf                ==6.4.0
python-dotenv        ==1.2.1
pytz                 ==2025.2
PyYAML               ==6.0.3
regex                ==2025.11.3
requests             ==2.32.5
requests-toolbelt    ==1.0.0
scipy                ==1.16.3
sniffio              ==1.3.1
tenacity             ==9.1.2
tiktoken             ==0.12.0
tqdm                 ==4.67.1
types-PyYAML         ==6.0.12.20250915
typing_extensions    ==4.15.0
typing-inspection    ==0.4.2
urllib3              ==2.5.0
xxhash               ==3.6.0
zstandard            ==0.25.0
pdfplumber           ==0.11.9